Un-deprecate `YieldProgress::noop()`.
After experience trying to avoid it, I have concluded that it should
continue being available. In particular, a library which never otherwise
constructs a `YieldProgress`, but only provides functions that accept
`YieldProgress`, should still be able to construct the noop instance for
its own tests or internal synchronous-call situations, without needing
to depend on anything more than the `YieldProgress` type (such as if it
is re-exported).