Add `split_concurrent()` for reporting concurrent tasks.
Note that we have to modify `ProgressFn` to allow non-`Send + Sync`
functions. This is coherent because the only case where it won't be
`Sync` is when the `sync` feature is enabled and we're internally using
`Rc` and so `YieldProgress` is itself not `Sync`. The public API still
requires `Send + Sync` for the function.