Replace all uses of `Rc` with `Arc`.
This removes an obstacle to having `Send + Sync` without `std`.
It does not change the public API at all.

It may in principle be slightly less efficient due to using atomic
operations even when not necessary. I consider this unlikely to be
significant in the kinds of application where `yield_progress` is
appropriate at all. (If you're calling `progress_without_yield()`
frequently, stop doing that.)